Stop linking requirements to tasks directly, currently done by searching
Summary:
Regarding the tasks not yet closed - We should avoid automatically creating
tasks for the requirements, as this is better done by the people planning the
work, as it happens. Also, there will normally be more than one task per
requirement, depending on how the work is organized. This means the people
should be fine with consulting directly Phabricator to find what's being worked
on at the moment.

Regarding the closed tasks - Those will result in a commit in the git
repository. As each git commit has a link to Phabricator, we can follow that to
list the relevant tasks.

Ref T1362

Test Plan: Check top-down and bottom-up reports

Reviewers: mb

Reviewed By: mb

Subscribers: reviews, cert, lj, dd

Maniphest Tasks: T1362

Differential Revision: https://p.daedalean.ai/D3150