Inject configuration and remove git module
The functionality of the git module is superseeded by the new repos
module which also takes care of managing multiple git repositories.

This change does not yet use the configuration other than for declaring
where documents and code are instead of looking through the certdoc and
codepath arguments (which have been removed).

In a later change, we will be able to remove as well the schema
argument, since it is also covered by the configuration.