Remove redundant Path and Repo from Req
Path will be replaced by a reference to the document where the
requirement was found, which contains the path if needed.

The Repo was already present in the struct as RepoName, so it was
duplicated.

This also removes the AddReq function which was quite useless in
reality and not used in actual code.