Ignore anonymous functions
In C++, some operators like operator[]() is confused with a lambda by
ctag, but actually parsed as both an anonymous function and an operator.

Since ctags produces 2 entries for the same element, we get later
conflicts when generating matrices and trying to order functions.

In addition, anonymous functions and methods are not relevant to
requirements.