Add assumption-specific attributes
Assumptions cannot be validated with the same attributes as any other
requirements. In order to handle them, the configuration now includes an
asmAttributes field that will be part of the document schema and will be
used to validate assumptions.

Assumptions also have implicit parents attribute declaration, and must
be any requirements with the requirement specification of the document
in which they are defined