Allow code with optional requirement links
In some cases we want some code to be able to link to requirements
without having to link to requirements. Such is the case of type
aliases, which is often used to link to requirements for type traits
and similar metaprogramming code, but does not always need to link to
requirements (it is also often used as simple type aliases for brevity).