Fix deduplication of code parents
This was done so that if any symbols are declared in multiple headers
and defined separately we don't have to link LLR's in all of them
manually.

However, it should take into account that the same symbol may have
different llrs depending on the document it belongs to (for instance for
implementation of a function declared by a third item).

As an examplee imagine that some dependency declares a function with no
definition. Then we have a document A which implements this function and
document B which also implements this function. Their parent tags must
necessarily be different, but the previous approach did not allow this.

Therefore, we only share the LLR across a single document (and in
reality they probably should be shared by binary only, but we don't have
enough information to do this).