Fix reqtraq resolution of control flow and data flow tags
The resolution was being invoked every time a new certdoc was added to
the graph, causing the issues to be duplicated once per document as new
documents get added.

Additionally, make sure that they get processed even if there are no
requirements in the document, and make sure that if we link an existing
DF and CF tag, which happens to belong to a different item than the one
the requirement belongs to, this is also flagged as an issue.